pasta/requirements.txt
# backend
Django
djangorestframework
drf-nested-routers

# frontend
Kivy
requests